pyproject.toml
[project]
name = "airflow-modern-data-lakehouse"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=3.1.0",
    "apache-airflow-providers-amazon>=9.15.0",
    "boto3-stubs==1.40.51",
    "cryptography>=46.0.2",
    "faker>=37.11.0",
    "pandas>=2.3.3",
    "pandas-stubs==2.3.2.250926",
    "polars>=1.34.0",
]
